Add scanning title provider
This provider, unlike the XML parsing provider, does not fail to
generate a title if the provided page is invalid xhtml. Instead,
it scans the text looking for <title> and </title> in order to
capture the content in between.

This option is also being made the default as it handles a wider
variety of real cases with regards to web content.